requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
pyodbc==4.0.39
python-dotenv==1.0.0
flask-login==0.6.2
geopy==2.4.1
python-socketio==5.9.0 